Refactored to module structure
